# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  common_models.FilterOperator:
    enum:
    - IS
    - IS_NOT
    - IS_ONE_OF_TERMS
    - IS_ONE_OF_TERMS_OR
    - IS_LESS_THAN_OR_EQUALS
    - IS_GREATER_THAN
    - EXIST
    - DOES_NOT_EXIST
    - NOT_CONTAINS
    - CONTAIN
    - DOES_NOT_CONTAIN
    - CONTAIN_ONE_OF
    - DOES_NOT_CONTAIN_ONE_OF
    - IS_ONE_OF
    - IS_NOT_ONE_OF
    - IS_BETWEEN
    - IS_NOT_BETWEEN
    - IS_IN_FIELDS
    - IS_NOT_IN_FIELDS
    - ENDS_WITH
    - NOT_ENDS_WITH
    - START_WITH
    - NOT_START_WITH
    type: string
    x-enum-varnames:
    - OpEquals
    - OpNotEquals
    - OpIn
    - OpInOr
    - OpLessOrEq
    - OpGreater
    - OpExists
    - OpNotExists
    - OpNotContains
    - OpContain
    - OpDoesNotContain
    - OpContainOneOf
    - OpDoesNotContainOneOf
    - OpIsOneOf
    - OpIsNotOneOf
    - OpIsBetween
    - OpIsNotBetween
    - OpIsInFields
    - OpIsNotInFields
    - OpEndsWith
    - OpNotEndsWith
    - OpStartWith
    - OpNotStartWith
  common_models.FilterType:
    properties:
      field:
        type: string
      operator:
        $ref: '#/definitions/common_models.FilterOperator'
      value: {}
    type: object
...
  dto.ActiveAlertTagRule:
    properties:
      appliedTagNames:
        items:
          type: string
        type: array
      conditions:
        items:
          type: integer
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.AlertTagResponse:
    properties:
      id:
        type: integer
      systemOwner:
        type: boolean
      tagColor:
        type: string
      tagName:
        type: string
    type: object
  dto.AlertTagRuleResponse:
    properties:
      active:
        type: boolean
      conditions:
        items:
          $ref: '#/definitions/common_models.FilterType'
        type: array
      createdBy:
        type: string
      createdDate:
        type: string
      deleted:
        type: boolean
      description:
        type: string
      id:
        type: integer
      lastModifiedBy:
        type: string
      lastModifiedDate:
        type: string
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.AlertTagRuleTagRef'
        type: array
    type: object
  dto.AlertTagRuleTagRef:
    properties:
      id:
        type: integer
      systemOwner:
        type: boolean
      tagColor:
        type: string
      tagName:
        type: string
    type: object
...
  dto.CreateAlertTagRequest:
    properties:
      tagColor:
        type: string
      tagName:
        type: string
    required:
    - tagName
    type: object
  dto.CreateAlertTagRuleRequest:
    properties:
      conditions:
        items:
          $ref: '#/definitions/common_models.FilterType'
        type: array
      description:
        type: string
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.AlertTagRuleTagRef'
        type: array
    required:
    - conditions
    - description
    - name
    - tags
    type: object
...
  dto.UpdateAlertTagRequest:
    properties:
      id:
        type: integer
      systemOwner:
        type: boolean
      tagColor:
        type: string
      tagName:
        type: string
    required:
    - id
    - tagName
    type: object
  dto.UpdateAlertTagRuleRequest:
    properties:
      conditions:
        items:
          $ref: '#/definitions/common_models.FilterType'
        type: array
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.AlertTagRuleTagRef'
        type: array
    required:
    - conditions
    - description
    - id
    - name
    - tags
    type: object
...
info:
  contact: {}
  description: UTMStack backend API — manages IAM, modules, flows, incidents, compliance
    and audit data.
  title: UTMStack Backend Service
  version: "1.0"
paths:
...
  /alert-tag-rules:
    get:
      parameters:
      - description: Filter by ID
        in: query
        name: id
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by condition field
        in: query
        name: conditionField
        type: string
      - description: Filter by condition value
        in: query
        name: conditionValue
        type: string
      - description: Filter by active status
        in: query
        name: ruleActive
        type: boolean
      - description: Filter by deleted status
        in: query
        name: ruleDeleted
        type: boolean
      - description: Comma-separated tag IDs
        in: query
        name: tagIds
        type: string
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Total items
              type: string
          schema:
            items:
              $ref: '#/definitions/dto.AlertTagRuleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alert tag rules
      tags:
      - Alert Tag Rules
    post:
      consumes:
      - application/json
      parameters:
      - description: Alert tag rule
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertTagRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertTagRuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create alert tag rule
      tags:
      - Alert Tag Rules
    put:
      consumes:
      - application/json
      parameters:
      - description: Alert tag rule
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertTagRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertTagRuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update alert tag rule
      tags:
      - Alert Tag Rules
  /alert-tag-rules/{id}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Deleted
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete alert tag rule
      tags:
      - Alert Tag Rules
    get:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertTagRuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert tag rule by ID
      tags:
      - Alert Tag Rules
  /alert-tag-rules/get-by-ids:
    get:
      parameters:
      - collectionFormat: csv
        description: Rule IDs (repeated param)
        in: query
        items:
          type: integer
        name: ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AlertTagRuleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert tag rules by IDs
      tags:
      - Alert Tag Rules
...
  /internal/alert-tag-rules/active:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ActiveAlertTagRule'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - InternalKey: []
      summary: List active alert tag rules (internal)
      tags:
      - Alert Tag Rules
...
  /utm-alert-tags:
    get:
      parameters:
      - description: Filter by tag name
        in: query
        name: tagName
        type: string
      - description: Filter by system owner
        in: query
        name: systemOwner
        type: boolean
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: Total items
              type: string
          schema:
            items:
              $ref: '#/definitions/dto.AlertTagResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alert tags
      tags:
      - Alert Tags
    post:
      consumes:
      - application/json
      parameters:
      - description: Alert tag
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AlertTagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create alert tag
      tags:
      - Alert Tags
    put:
      consumes:
      - application/json
      parameters:
      - description: Alert tag
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertTagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update alert tag
      tags:
      - Alert Tags
  /utm-alert-tags/{id}:
    delete:
      parameters:
      - description: Alert tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Deleted
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete alert tag
      tags:
      - Alert Tags
    get:
      parameters:
      - description: Alert tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AlertTagResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert tag by ID
      tags:
      - Alert Tags
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"